// src/types/Address.ts
enum Address {
    ETHxp = "ETHxp",
    ETH = "ETH",
    fDAIx = "0xF2d68898557cCb2Cf4C10c3Ef2B034b2a69DAD00",
    fDAI = "0x88271d333C72e51516B67f5567c728E702b3eeE8",
    fUSDCx = "0x8aE68021f6170E5a766bE613cEA0d75236ECCa9a",
    fUSDC = "0xc94dd466416A7dFE166aB2cF916D3875C049EBB7",
    fAMKT = "0x123",
    fAMKTxp = "0x321",
    unlisted = "unlisted",
}

export default Address;